% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_county_data.R
\name{assemble_county_data}
\alias{assemble_county_data}
\title{Assemble sensor string data from different deployments}
\usage{
assemble_county_data(input_path = NULL, county = "")
}
\arguments{
\item{input_path}{Path to the *.csv files to be assembled. Default is the
Open Data/Deployment Data folder on the CMAR Operations shared drive (user
must be connected to the Perennia VPN).}

\item{county}{A character string. If \code{input_path = NULL}, \code{county}
is the next folder on the path.}
}
\value{
Returns a dataframe of the assembled county data.
}
\description{
Only users connected to the CMAR Operations shared drive can use
  the default \code{input_path}.
}
\details{
Right now can only handle files exported from
  \code{format_for_opendata()}. Future versions will also be able to assemble
  data from \code{convert_to_tidydata()}.
}
\seealso{
Other OpenData CMAR: 
\code{\link{export_county_files}()},
\code{\link{import_strings_data}()}
}
\author{
Danielle Dempsey
}
\concept{OpenData CMAR}
