% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cp.R
\name{calculate_cp}
\alias{calculate_cp}
\title{Calculate the solubility of dissolved oxygen}
\usage{
calculate_cp(dat.wide, Sal = NULL, P_atm = NULL)
}
\arguments{
\item{dat.wide}{Dataframe with at least one column: \code{Temperature}.
Corresponding salinity (psu) and pressure (atm) data may be included in
columns \code{Salinity} and \code{Pressure}.}

\item{Sal}{A single value of salinity (ppt). This value should be specified
if there is no \code{Salinity} column in \code{dat.wide}. Default is
\code{Sal = NULL}. Note: if \code{Sal} is specified when there is a
\code{Salinity} column in \code{dat.wide}, function will stop with an
error.}

\item{P_atm}{A single value of barometric pressure (atm). This value should
be specified if there is no \code{Pressure} column in \code{dat.wide}.
Default is \code{P_atm = NULL}. Note: if \code{P_atm} is specified when
there is a \code{Pressure} column in \code{dat.wide}, function will stop
with an error.}
}
\description{
Calculate the solubility of dissolved oxygen based on
  temperature, salinity and barometric pressure, following the equations in
  Benson and Krause 1984.
}
\details{
Solubility of dissolved oxygen (mg/L) is calculated using the
  equations in Benson and Krause 1984.

  These equations should only be used when temperature is between 0 and 40
  degrees-Celsius and salinity is between 0 and 40 psu.

  Results from this function should match those from the USGS DOTABLES at
  \url{https://water.usgs.gov/water-resources/software/DOTABLES/}.

  For more info see equations 24 and 32, and Table 2 from Benson and Krause
  1984.

  Benson, Bruce B., Krause, Daniel, (1984), The concentration and isotopic
  fractionation of oxygen dissolved in freshwater and seawater in equilibrium
  with the atmosphere, Limnology and Oceanography, 3, doi:
  10.4319/lo.1984.29.3.0620.
}
