% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_degree_days.R
\name{calculate_degree_days}
\alias{calculate_degree_days}
\title{Converts temperature data to degree-days}
\usage{
calculate_degree_days(
  dat.tidy,
  start.date,
  end.date,
  depths.to.include = "ALL",
  group.by.depth = TRUE
)
}
\arguments{
\item{dat.tidy}{Data in tidy format, as returned by the function
\code{convert_to_tidydata()}. Must include three columns: \code{DATE}
(POSIXct), \code{VARIABLE} (character), and \code{VALUE} (numeric). May also
include a column \code{DEPTH} (factor), which can be used to filter the data
included in the degree-days calculation. Other columns wil be ignored.}

\item{start.date}{First day that should be included in the calculation. If
\code{start.date} does not include a time, then the start time is assumed to
be midnight. Accepted orders for \code{start.date} are: "ymd IMS p", "Ymd
IMS p", "Ymd HM", "Ymd HMS", "dmY HM", "dmY HMS", "Ymd", "ymd".}

\item{end.date}{Last day that should be included in the calculation. If
\code{end.date} does not include a time, then the end time is assumed to be
"23:59:59". Accepted orders for \code{start.date} are: "ymd IMS p", "Ymd IMS
p", "Ymd HM", "Ymd HMS", "dmY HM", "dmY HMS", "Ymd", "ymd".}

\item{depths.to.include}{Numeric vector of depths to include in the
calculation. Default is \code{depths.to.include = "ALL"}, which includes all
depths in the \code{DEPTH} column of \code{dat.tidy}.}

\item{group.by.depth}{Logical value. By default, \code{group.by.depth = TRUE},
and so the degree days are calculated and returned for each depth specified
in \code{depths.to.include}. If \code{group.by.depth = FALSE} the
degree-days will averaged over all depths.}
}
\value{
If \code{group.by.depth = FALSE}, returns the calue of the calculated
 degree-days averaged over all depths specified in \code{depths.to.include}.
 If \code{group.by.depth = TRUE}, returns a tibble with two columns:
 \code{DEPTH}, and the corresponding \code{Degree_Days}.
}
\description{
Converts temperature data to degree-days
}
\seealso{
Other calculate: 
\code{\link{convert_DO_percent_to_concentration}()}
}
\author{
Danielle Dempsey
}
\concept{calculate}
