% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_deployment_logs.R
\name{combine_deployment_logs}
\alias{combine_deployment_logs}
\title{Combines deloyment logs from a station to make deployment history}
\usage{
combine_deployment_logs(path)
}
\arguments{
\item{path}{Path to station folder.}
}
\value{
Fill this in
}
\description{
Combines deloyment logs from a station to make deployment history
}
\seealso{
Other format: 
\code{\link{convert_depth_to_ordered_factor}()},
\code{\link{convert_to_tidydata}()},
\code{\link{trim_data}()}
}
\author{
Danielle Dempsey
}
\concept{format}
