% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_all_data.R
\name{compile_all_data}
\alias{compile_all_data}
\title{Compiles HOBO, aquaMeasure, and Vemco data from a single deployment}
\usage{
compile_all_data(
  path,
  area.name,
  deployment.range,
  trim,
  serial.table.HOBO = NULL,
  file.type = NULL,
  serial.table.aM = NULL,
  depth.vemco = NULL
)
}
\arguments{
\item{path}{File path to the Hobo, aquaMeasure, and/or Vemco folders.}

\item{area.name}{Area where the HOBO was deployed.}

\item{deployment.range}{The start and end dates of deployment from the
deployment log. Must be in format "2018-Nov-15 to 2020-Jan-24".}

\item{trim}{Logical value indicating whether to trim the data to the dates
specified in \code{deployment.range}. (Note: four hours are added to the
retrieval date to account for AST, e.g., in case the sensor was retrieved
after 20:00 AST, which is 00:00 UTC the next day.) Default is \code{trim =
TRUE}.}

\item{serial.table.HOBO}{A table with the serial number of each HOBO sensor on the
string, in the form "HOBO-xxxxxxxx" (first column) and corresponding
variable it measured at the depth it was deployed in the form
"Temperature-2m" (second column).}

\item{file.type}{Character string indicating whether the HOBO data is in .csv
or .xlsx format. All data files that are being compiled must have the same
file extension. Note that the HOBO software exports csv files in true UTC,
but xlsx files account for daylight savings time. This function will
automatically convert the datetimes in xlsx files to true UTC. Default is
\code{file.type = "csv"}. Alternative is \code{file.type = "xlsx"}.}

\item{serial.table.aM}{A table with the serial number of each aquaMeasure on the
string, in the form "aquaMeasure-xxxxxx" (first column; note the capital
"M") and its corresponding depth in the form "2m" (second column).}

\item{depth.vemco}{Character string describing the depth at which the Vemco
was deployed, in the format "10m".}
}
\value{
Returns a dataframe or exports a spreadsheet with the formatted Vemco
 data in two columns: the timestamp (UTC, in the format "Y-m-d H:M:S") and
 temperature value (degree celsius, rounded to three decimal places).
 Metadata at the top of each column indicates the deployment range, the
 sensor serial number, and the depth of the sensor. Each datetime column
 shows the timezone as extracted from the Vue software.

 To include the metadata, all values were converted to class
 \code{character}. To manipulate the data, the values must be converted to
 the appropriate class (e.g., \code{POSIXct} for the datetimes and
 \code{numeric} for temperature values). This can be done using the function
 \code{convert_to_tidydata()}.
}
\description{
Compiles HOBO, aquaMeasure, and Vemco data from a single deployment
}
\details{
Write that it called the other functions and to look at those for help
}
\seealso{
Other compile: 
\code{\link{compile_HOBO_data}()},
\code{\link{compile_aquaMeasure_data}()},
\code{\link{compile_vemco_data}()},
\code{\link{convert_to_tidydata}()}
}
\author{
Danielle Dempsey
}
\concept{compile}
