% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_HOBO_data.R
\name{compile_HOBO_data}
\alias{compile_HOBO_data}
\title{Compiles temperature data from HOBO deployment}
\usage{
compile_HOBO_data(path.HOBO, area.name, serial.table, deployment.range)
}
\arguments{
\item{path.HOBO}{File path to the HOBO data. Should end with \code{/Hobo}.}

\item{area.name}{Area where HOBO was deployed.}

\item{serial.table}{A table with the serial number of each HOBO (first column) and
corresponding depth at which it was deployed (second column).}

\item{deployment.range}{The start and end dates of deployment from the
deployment log. Must be in format "2018-Nov-15 to 2020-Jan-24".}
}
\value{
Returns a dataframe and exports a spreadsheet with the HOBO
 temperature data, including the appropriate metadata. Note that to include
 the metadata, all values were converted to class \code{character}. To manipulate the data, the
 values must be converted to the appropriate class (e.g., \code{POSIXct} for \code{DATE},
 \code{numeric} for temperature values). This can be done using the function \code{convert_to_tidydata()}.
}
\description{
This function compiles the data from a HOBO deployment at
 different depths into a single spreadsheet.
}
\details{
I haven't done anything with the DATES. Still need to un-account for
 daylight savings. To be discussed.
}
\seealso{
Other compile: 
\code{\link{compile_aquaMeasure_data}()}
}
\author{
Danielle Dempsey
}
\concept{compile}
