% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_Vemco_data.R
\name{compile_vemco_data}
\alias{compile_vemco_data}
\title{Formats temperature data from Vemco deployment}
\usage{
compile_vemco_data(
  path.vemco,
  area.name,
  depth.vemco,
  deployment.range,
  trim = TRUE,
  export.csv = FALSE
)
}
\arguments{
\item{path.vemco}{File path to the Vemco folder. This folder should have one
csv file that was extracted using Vue software. Other file types in the
folder will be ignored.}

\item{area.name}{Area where the Vemco was deployed.}

\item{depth.vemco}{Depth at which the Vemco was deployed.}

\item{deployment.range}{The start and end dates of deployment from the
deployment log. Must be in format "2018-Nov-15 to 2020-Jan-24".}

\item{trim}{Logical value indicating whether to trim the data to the dates
specified in \code{deployment.range}. (Note: four hours are added to the
retrieval date to account for AST, e.g., in case the sensor was retrieved
after 20:00 AST, which is 00:00 UTC the next day.) Default is \code{trim =
TRUE}.}

\item{export.csv}{Logical value indicating whether to export the compiled data
as a .csv file. If \code{export.csv = TRUE}, the compiled data will not be
returned to the global environment. Default is \code{export.csv = FALSE}.}
}
\value{
Returns a dataframe or spreadsheet with the formatted Vemco datain two
 columns: the timestamp (UTC, in the format "Y-m-d H:M:S") and temperature
 value (degree celsius, rounded to three decimal places). Metadata at the top
 of each column indicates the deployment range, the sensor serial number, and
 the depth of the sensor. Each datetime column shows the timezone as
 extracted from the Vue software.

 To include the metadata, all values were converted to class
 \code{character}. To manipulate the data, the values must be converted to
 the appropriate class (e.g., \code{POSIXct} for the datetimes and
 \code{numeric} for temperature values). This can be done using the function
 \code{convert_to_tidydata()}.
}
\description{
This function formats data from a Vemco deployment so it can be
 compiled with the HOBO and aquaMeasure data.
}
\seealso{
Other compile: 
\code{\link{compile_HOBO_data}()},
\code{\link{compile_aquaMeasure_data}()},
\code{\link{convert_to_tidydata}()}
}
\author{
Danielle Dempsey
}
\concept{compile}
