% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_DEPTH_to_ordered_factor.R
\name{convert_depth_to_ordered_factor}
\alias{convert_depth_to_ordered_factor}
\title{Converts values in the DEPTH column to an ordered factor}
\usage{
convert_depth_to_ordered_factor(dat.tidy)
}
\arguments{
\item{dat.tidy}{Data in tidy format, as returned by the function
\code{convert_to_tidydata()}. Must include at least one column: \code{DEPTH}
(factor, character, or numeric). The values of \code{DEPTH} must be
coercible to class \code{numeric}.}
}
\value{
Returns dat.tidy, with the DEPTH column converted to an ordered
 factor. The smallest value will be assigned to the first level, the second
 smallest value will be assigned to the second level, and so on.
}
\description{
This function simplifies converting the DEPTH column from class
 \code{numeric}, \code{factor}, or \code{character} to an ordered factor,
 which is preferred for using the function \code{plot_variables_at_depth()}.
}
\details{
To use the function \code{plot_variables_at_depth()}, \code{DEPTH}
 must be a factor because \code{DEPTH} is mapped to a discrete color scale.
 For the legend to display correctly (e.g. depths arranged from shallow to
 deep), \code{DEPTH} must be an ordered factor.
}
\seealso{
Other format: 
\code{\link{convert_DO_percent_to_concentration}()},
\code{\link{convert_to_tidydata}()},
\code{\link{format_for_opendata}()},
\code{\link{trim_data}()}
}
\author{
Danielle Dempsey
}
\concept{format}
