% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_deployment_log.R
\name{create_deployment_log}
\alias{create_deployment_log}
\title{Writes deployment log from the NSDFA tracking sheet}
\usage{
create_deployment_log(
  path.tracking.sheet,
  path.export,
  station,
  deployment.date
)
}
\arguments{
\item{path.tracking.sheet}{Path to where "2020-08-10 - NSDFA_Tracking.xlsx" is
saved.}

\item{path.export}{Path to the Log folder.}

\item{station}{Station name}

\item{deployment.date}{Date of deployment as a character string in the format
"\%Y-\%m-\%d"}
}
\value{
Fill this out
}
\description{
Writes deployment log from the NSDFA tracking sheet
}
\details{
I tried to match Recv_Method from the tracking sheet with Mount_Type
 in the Log, but it doesn't match existing logs.
}
\author{
Danielle Dempsey
}
