% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_deployment_log.R
\name{create_initial_deployment_log}
\alias{create_initial_deployment_log}
\title{Exports an initial deployment log}
\usage{
create_initial_deployment_log(
  station,
  waterbody = "lookup",
  path.export,
  n.sensors = 5
)
}
\arguments{
\item{station}{Station name (for the Location_Description column).}

\item{waterbody}{Waterbody name (for the Deployment_Waterbody column). If
\code{waterbody} == "lookup" (the default), the function will look up the
waterbody corresponding to \code{station} in the Area Info tab of the
STRINGS TRACKING google spreadsheet.}

\item{path.export}{Path to where the log should be exported.}

\item{n.sensors}{Number of sensors that will be included in the string, if
known. Default is \code{n.sensors = 5}.}
}
\value{
Returns deployment log in .xlsx format.
}
\description{
Exports an initial deployment log
}
\details{
Deployment_Waterbody and Location_Description are filled in.
  Remaining columns are empty.
}
\author{
Danielle Dempsey
}
