% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_deployment_log.R
\name{create_log_from_NSDFA_tracking}
\alias{create_log_from_NSDFA_tracking}
\title{Writes deployment log from the NSDFA tracking sheet}
\usage{
create_log_from_NSDFA_tracking(
  path.tracking.sheet,
  temp.sheet = "TempMetaData",
  path.export,
  station,
  deployment.date
)
}
\arguments{
\item{path.tracking.sheet}{Full path to the NSDFA tracking sheet (including
the file name and extension). This must be a .xlsx file.}

\item{temp.sheet}{Name of the tab with the temperature data (in "").}

\item{path.export}{Path to the Log folder.}

\item{station}{Station name.}

\item{deployment.date}{Date of deployment as a character string in the format
"Y-m-d"}
}
\value{
Returns deployment log in .xlsx format.
}
\description{
Writes deployment log from the NSDFA tracking sheet
}
\details{
Imports the NSDFA tracking sheet, filters for the station and date of
 interest, re-formats into the deployment log format, and exports to the Log
 folder.

 DD: I tried to match Recv_Method from the tracking sheet with Mount_Type in
 the Log, but it doesn't match existing logs.
}
\author{
Danielle Dempsey
}
