% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_log.R
\docType{data}
\name{deployment_log}
\alias{deployment_log}
\title{Example log}
\format{
This example log illustrates the columns required for use with the
  `read_deployment_log()` function. The deployment log must be saved in the
  Log folder, be named "station deployment- date Log", and be in .csv, .xlsx,
  or .xls format. There is one row for each sensor on the string, and 10
  columns (other columns will be ignored by the package):  \describe{
  \item{Deployment_Waterbody}{Waterbody where string was deployed.}
  \item{Location_Description}{The station name.}
  \item{Lease#}{If located on an aquaculture site, the lease number (NA otherwise).}
  \item{Deployment}{The deployment date, in the order "Ymd".}
  \item{Retrieval}{The retrieval date, in the order "Ymd".}
  \item{Logger_Latitude}{The latitude at which the string was deployed.}
  \item{Logger_Longitude}{The longitude at which the string was deployed.}
   \item{Logger_Model}{The type of sensor.}
   \item{Serial#}{The sensor serial number.}
   \item{Sensor_Depth}{The depth at which the sensor was deployed.}

  }
}
\source{
Data from Coastal Monitoring Program
}
\usage{
deployment_log
}
\description{
Example log
}
\keyword{datasets}
