% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_correction_factors.R
\name{DO_pressure_correction}
\alias{DO_pressure_correction}
\title{Pressure correction factor for dissolved oxygen measurements}
\usage{
DO_pressure_correction(dat.wide, P_atm = NULL)
}
\arguments{
\item{dat.wide}{Dataframe with at least one column: \code{Temperature}.
Corresponding presure (atm) data may be included in column
\code{Pressure}.}

\item{P_atm}{A single value of barometric pressure (atm). This value should
be specified if there is no \code{Pressure} column in \code{dat.wide}.
Default is \code{P_atm = NULL}. Note: if \code{P_atm} is specified when
there is a \code{Pressure} column in \code{dat.wide}, function will stop
with an error.}
}
\value{
Returns \code{dat.wide} with an additional column, \code{F_p}.
}
\description{
Pressure correction factor for dissolved oxygen measurements
}
\details{
Dissolved oxygen concentration (and partial pressure?) measurements
  should be corrected for atmospheric pressure if the pressure deviates from
  substantially from 1 atm (Benson & Krause, 1984).

  This function calculates pressure correction factor following Benson and
  Krause (1984), as suggested by USGS (2011), similar to what is described in
  the HOBO manual HOBO U26 Percent Saturation Calculation.pdf (except this
  function account for salinity in the water vapour calculation).

  Equation 24 in Benson & Krause 1984:

  \deqn{C_{p} = C_star x Pressure x (((1 - P_{wv} / Pressure) x (1 - theta x
  Pressure)) / ((1 - P_{wv}) x (1 - theta)))}

  \deqn{F_{p} = Pressure x (((1 - P_{wv} / Pressure) x (1 - theta x
  Pressure)) / ((1 - P_{wv}) x (1 - theta)))}

  \eqn{P_{wv}} is water vapour pressure, which depends on temperature and
  salinity and \eqn{theta} depends on the second virial coefficient of
  oxygen, and is calculated using temperature (see Table 1 of Benson and
  Krause 1984).

  \bold{References}

  Benson, Bruce B., Krause, Daniel, (1984), \emph{The concentration and
  isotopic fractionation of oxygen dissolved in freshwater and seawater in
  equilibrium with the atmosphere, Limnology and Oceanography}, 3, doi:
  10.4319/lo.1984.29.3.0620.

  USGS. \emph{Change to Solubility Equations for Oxygen in Water.} Technical
  Memorandum 2011.03. USGS Office of Water Quality, 2011.
}
