% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_correction_factors.R
\name{DO_salinity_correction}
\alias{DO_salinity_correction}
\title{Salinity correction factor for dissolved oxygen measurements}
\usage{
DO_salinity_correction(dat.wide, Sal = NULL, method = "garcia-gordon")
}
\arguments{
\item{dat.wide}{Dataframe with at least one column: \code{Temperature}.
Corresponding salinity (psu) data may be included in column
\code{Salinity}.}

\item{Sal}{A single value of salinity (psu). This value must be specified if
there is no \code{Salinity} column in \code{dat.wide}. Default is \code{Sal
= NULL}. Note: if \code{Sal} is specified when there is a \code{Salinity}
column in \code{dat.wide}, the function will stop with an error.}

\item{method}{Equation to use to calculate dissolved oxygen solubility.
Options are \code{"garcia-gordon"} (the default) and \code{benson-krause}.}
}
\value{
Returns \code{dat.wide} with an additional column, \code{F_s}.
}
\description{
Salinity correction factor for dissolved oxygen measurements
}
\details{
Dissolved oxygen concentration measured by HOBO sensors needs to be
  corrected for salinity (see manual:
  \url{https://www.onsetcomp.com/files/manual_pdfs/15603-B-MAN-U26x.pdf}).

  The salinity correction factor can be calculated using the Benson and
  Krause 1984 equations or the Garcia and Gordon 1992 equations. These
  equations should only be used when 0 < Temperature < 40 degrees Celcius, 0
  < Salinity < 40 PSU and 0.5 < Pressue < 1.1 atm.

  \bold{Benson & Krause, 1984}:

  Used by the USGS DOTables (USGS, 2011).

  Results from this equation should match those from the USGS DOTABLES (part
  C) at \url{https://water.usgs.gov/water-resources/software/DOTABLES/}.

  Final term in Equation 32:

  \deqn{F_{s} = exp(-Salinity * (0.017674 - 10.754 / T_{Kelvin} + 2140.7 /
  T_{Kelvin}^2))}

  \bold{Garcia &  Gordon, 1992:}

  Garcia & Gordon re-fit the Benson & Krause data with a higher order
  polynomial and defined a scaled temperature (\eqn{T_s}).

  This correction factor is used in the SCOR WG 142 (Part C).

  Results from this equation are very similar to the Benson & Krause
  correction factor (to ~4 decimal places).

  \eqn{F_{s} = exp(Salinity x (-6.24523E-3 -7.37614E-3 x T_s -1.03410E-2 x
  T_s^2 -8.17083E-3 x T_s^3) -4.88682E-7 * Salinity^2)}

  Note: the HOBO Dissolved Oxygen Assistant Software salinity correction
  factor uses the same form as the Garcia and Gordon equation, but different
  coefficients (negligible difference in the resulting \deqn{F_s}).

  \bold{References}

  Benson, Bruce B., Krause, Daniel, (1984), \emph{The concentration and
  isotopic fractionation of oxygen dissolved in freshwater and seawater in
  equilibrium with the atmosphere, Limnology and Oceanography}, 3, doi:
  10.4319/lo.1984.29.3.0620.

  Bittig, H. &.-J. (2016). \emph{SCOR WG 142: Quality Control Procedures for
  Oxygen and Other Biogeochemical Sensors on Floats and Gliders.
  Recommendations on the conversion between oxygen quantities for Bio-Argo
  floats and other autonomous sensor platforms.}
  \url{https://repository.oceanbestpractices.org/handle/11329/417}.

  Garcia, H., and L. Gordon (1992), \emph{Oxygen solubility in seawater:
  Better fitting equations}, Limnol. Oceanogr., 37(6).

  USGS. \emph{Change to Solubility Equations for Oxygen in Water.} Technical
  Memorandum 2011.03. USGS Office of Water Quality, 2011.
}
