% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_county_files.R
\name{export_county_files}
\alias{export_county_files}
\title{Export data into a csv and/or rds file}
\usage{
export_county_files(
  dat,
  file_name = "",
  output_path = NULL,
  export_csv = TRUE,
  export_rds = TRUE,
  return_global = FALSE
)
}
\arguments{
\item{dat}{Dataframe to be exported, e.g., output from
\code{export_county_data()}.}

\item{file_name}{Name for the output file. Will be appended with today's
date.}

\item{output_path}{Path to where assembled file(s) should be exported. The
default will export the csv file to the Open Data/Submissions folder on the
CMAR Operations shared drive (for submission to the Open Data Portal) and
the rds file to the Open Data/County Datasets folder. For custom
\code{output_path} entries, the csv and rds files will be exported to the
same folder.}

\item{export_csv}{Logical argument indicating whether the data should be
exported as a *.csv file. File name will be county_todays-date.csv. Default
is \code{TRUE}. Note: \code{TIMESTAMP} is converted to a character before
exporting to remove UTC formatting (2018-12-23T05:00:00Z). When re-imported
into R, the UTC timezone should be added using the \code{force_tz()}
function from the lubridate package.}

\item{export_rds}{Logical argument indicating whether the assembled data
should be exported as a *.rds file. File name will be
county_todays-date.rds. Default is \code{TRUE}.}

\item{return_global}{Logical argument indicating whether the assembled data
should be returned to the global environment. Default is \code{FALSE}.}
}
\value{
Exports \code{dat} as a csv and/or rds file.
}
\description{
Export data into a csv and/or rds file
}
\seealso{
Other OpenData CMAR: 
\code{\link{assemble_county_data}()},
\code{\link{import_strings_data}()}
}
\author{
Danielle Dempsey
}
\concept{OpenData CMAR}
