% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_opendata.R
\name{export_opendata}
\alias{export_opendata}
\title{Export sensor string data for the OpenData portal}
\usage{
export_opendata(path.data, path.export, deployment.name, file.name)
}
\arguments{
\item{path.data}{Path to deployment folder}

\item{path.export}{Path to where csv will be exported}

\item{deployment.name}{Name of deployment folder. Will be used as name of the csv file.}

\item{file.name}{Name of file to import. Must be a csv}
}
\value{
Exports a csv file in the format for the OpenData portal.
}
\description{
Export sensor string data for the OpenData portal
}
\author{
Danielle Dempsey
}
\concept{OpenData}
