% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_days.R
\name{filter_days}
\alias{filter_days}
\title{Filters out data from days when the value exceeds a threshold}
\usage{
filter_days(dat.tidy, threshold = 20, min.exceedence = 1, n.days.to.filter = 1)
}
\arguments{
\item{dat.tidy}{Dataframe with at least three columns: \code{TIMESTAMP} (must
be possible to convert to a Date object), \code{DEPTH}, and \code{VALUE}.
If column \code{VARIABLE} is included, it must have one unique entry. Other
columns will be ignored.}

\item{threshold}{The threshold to trigger filtering data (inclusive. Default
is \code{threshold = 20}.}

\item{min.exceedence}{The minimum number of observations in a day to trigger
filtering data (inclusive). Default is \code{min.exceedence = 1}.}

\item{n.days.to.filter}{Number of days to filter out when \code{threshold}
and \code{min.exceedence} are exceeded. If \code{n.days.to.filter = 1} (the
default), then all values from the day the threshold is exceeded will be
filtered out. If \code{n.days.to.filter = 2}, then all values from the day
the threshold is exceeded and the following day will be filtered out, etc.}
}
\value{
Returns dat.tidy, filtered when triggered by \code{threshold} and
  \code{min.exceedence}.
}
\description{
Filters out all data from days when the value exceeds a given
  threshold for a given number of observations. Option to filter out
  additional days after the threshold is exceeded.

  \code{dat.tidy } should only include data for one variable.
}
