% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_deployment_log.R
\name{format_deployment_log}
\alias{format_deployment_log}
\title{Create formatted xlsx workbook add log information}
\usage{
format_deployment_log(deployment_log)
}
\arguments{
\item{deployment_log}{Deployment log created from NSDFA tracking sheet or}
}
\value{
Returns deployment log in .csv format.
}
\description{
Create formatted xlsx workbook add log information
}
\details{
Called by \code{create_log_from_NSDFA_tracking()} and
  \code{create_initial_deployment_log()}.
}
\author{
Danielle Dempsey
}
