% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_for_opendata.R
\name{format_for_opendata}
\alias{format_for_opendata}
\title{Format sensor string data for the OpenData portal}
\usage{
format_for_opendata(dat.tidy, lat.deploy, long.deploy, waterbody, county)
}
\arguments{
\item{dat.tidy}{Data from a single deployment in tidy format, as returned by
the function \code{convert_to_tidydata()}. Must include four columns:
\code{DATE} (POSIXct), \code{VARIABLE} (character), \code{DEPTH} (ordered
factor), and \code{VALUE} (numeric).}

\item{lat.deploy}{The latitude at which the station was deployed.}

\item{long.deploy}{The longitude at which the station was deployed.}

\item{waterbody}{The water body in which the station was deployed.}

\item{county}{The county in which the station was deployed.}
}
\value{
Returns a dataframe of tidy data in the format for the OpenData portal for a single deployment,
}
\description{
Format sensor string data for the OpenData portal
}
\seealso{
Other format: 
\code{\link{convert_to_tidydata}()}
}
\author{
Danielle Dempsey
}
\concept{format}
