% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot.R
\name{get_colour_palette}
\alias{get_colour_palette}
\title{Returns nice colour palette based on the number of unique DEPTH values}
\usage{
get_colour_palette(dat.tidy)
}
\arguments{
\item{dat.tidy}{Data to be plotted, in tidy format, as returned by the
function \code{convert_to_tidydata()}. Must include the column
\code{DEPTH}.}
}
\value{
Returns a vector of hex colours from the viridis palette
}
\description{
Returns nice colour palette based on the number of unique DEPTH values
}
\details{
If there are 6 or less unique values of \code{DEPTH}, returns a
 discrete viridis colour palette with 6 colours (Option D). If there are more
 than 6 unique values of \code{DEPTH}, returns a discrete colour palette with
 the number of colours equal to the number of unique values of \code{DEPTH}.
}
\seealso{
Other plot: 
\code{\link{get_xaxis_breaks}()},
\code{\link{plot_variables_at_depth}()}
}
\author{
Danielle Dempsey
}
\concept{plot}
