% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{get_xaxis_breaks}
\alias{get_xaxis_breaks}
\title{Returns nice major and minor breaks and label format based on timespan of the data}
\usage{
get_xaxis_breaks(timespan)
}
\arguments{
\item{timespan}{Number of days of data to be plotted.}
}
\value{
Returns a dataframe with 1 observation of 3 variables
 \code{date.breaks.major}, \code{date.breaks.minor},
 \code{date.labels.format}.
}
\description{
Returns nice major and minor breaks and label format based on timespan of the data
}
\details{
This is a temporary function to help prepare OpenData reports.
}
\seealso{
Other OpenData: 
\code{\link{write_report_table}()}
}
\author{
Danielle Dempsey
}
\concept{OpenData}
