% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_strings_data.R
\name{import_strings_data}
\alias{import_strings_data}
\title{Import string data}
\usage{
import_strings_data(
  input_path = NULL,
  output_path = "",
  county = "ALL",
  add_county_col = TRUE,
  export_csv = FALSE,
  return_global = TRUE
)
}
\arguments{
\item{input_path}{Placeholder}

\item{output_path}{Path to where assembled file should be exported}

\item{county}{Optional filter}

\item{export_csv}{Logical argument indicating whether the data
should be exported as a *.csv file.}

\item{return_global}{Logical argument indicating whether the assembled data
should be returned to the global environment. Default is \code{FALSE}.}
}
\description{
Import string data
}
\details{
More about the path here
}
\seealso{
Other OpenData: 
\code{\link{assemble_opendata_submission}()},
\code{\link{write_report_table}()}
}
\author{
Danielle Dempsey
}
\concept{OpenData}
