% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temp_DO_sal.R
\name{plot_temp_DO_sal}
\alias{plot_temp_DO_sal}
\title{Exports ggplot2 objects for temperauture, dissolved oxygen, and
 salinity on separate panes}
\usage{
plot_temp_DO_sal(
  dat.tidy,
  vars.to.plot = c("Temperature", "Dissolved Oxygen", "Salinity"),
  color.palette = rev(viridis(6, option = "D")),
  date.breaks.major = "2 month",
  date.breaks.minor = "1 month"
)
}
\arguments{
\item{dat.tidy}{Data in tidy format, as returned by the function
\code{convert_to_tidydata()}. Must include four columns: \code{DATE}
(POSIXct), \code{VALUE}, (numeric), \code{VARIABLE} (character),
\code{DEPTH} (ordered factor).}

\item{vars.to.plot}{A character vector of variables for which to return plots.
Options are any combination of "Temperature", "Dissolved Oxygen", and
"Salinity". Default is \code{vars.to.plot = c("Temperature", "Dissolved
Oxygen", "Salinity")}.}

\item{color.palette}{Color palette of hex colors onto which \code{DEPTH} will
be mapped. Default is \code{color.palette = rev(viridis(6, option = "D"))}.}

\item{date.breaks.major}{Intervals for major breaks. Default is
\code{date.breaks.major = "2 month"}.}

\item{date.breaks.minor}{Intervals for minor breaks. Default is
\code{date.breaks.minor = "1 month"}.}
}
\value{
Returns a list of ggplot2 plots of Temperature, Dissolved Oxygen, and/or Salinity.
}
\description{
Exports ggplot2 objects for temperauture, dissolved oxygen, and
 salinity on separate panes
}
\author{
Danielle Dempsey
}
\concept{plot}
