% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variables_at_depth.R
\name{plot_variables_at_depth}
\alias{plot_variables_at_depth}
\title{Exports ggplot2 object(s) of variables at depth over time}
\usage{
plot_variables_at_depth(
  dat.tidy,
  plot.title = "",
  color.palette = NULL,
  date.breaks.major = NULL,
  date.breaks.minor = NULL,
  date.labels.format = NULL,
  standard.DO.ylims = TRUE,
  alpha.points = 1,
  legend.name = "Depth (m)",
  legend.position = "right"
)
}
\arguments{
\item{dat.tidy}{Data in tidy format, as returned by the function
\code{convert_to_tidydata()}. Must include four columns: \code{TIMESTAMP}
(POSIXct), \code{VARIABLE} (character), \code{DEPTH} (ordered factor), and
\code{VALUE} (numeric).}

\item{plot.title}{Title for plot. Default is no title.}

\item{color.palette}{Color palette of hex colors onto which \code{DEPTH} will
be mapped. Required if there are more than 6 levels in \code{DEPTH}.
Default is \code{color.palette = rev(viridis(6, option = "D"))}.}

\item{date.breaks.major}{Intervals for major breaks. Default is selected by
\code{get_xaxis_breaks()}.}

\item{date.breaks.minor}{Intervals for minor breaks. Default is
\code{date.breaks.minor = "1 month"}.}

\item{date.labels.format}{Format for the date labels. Default is "\%y-\%b"
(two-digit year-three-letter month).}

\item{standard.DO.ylims}{If \code{TRUE}, the y-limits for dissolved oxygen
are set to c(60, 130) \%. If \code{FALSE}, the y-limits are set to the
\code{ggplot} default.}

\item{alpha.points}{Value indicating the transparency of the points. 0 is
most transparent; 1 is opaque.}

\item{legend.name}{Name for the legend. Must be a character string. Default
is \code{legend.name = "Depth (m)"}.}

\item{legend.position}{Position for the legend. Passed to \code{ggpubr}.
Default is \code{legend.position = "right"}.}
}
\value{
Returns ggplot2 object, a single figure with the plots for each
  variable in \code{tidy.data} stacked in a column is returned.
}
\description{
Exports ggplot2 object(s) of variables at depth over time
}
\seealso{
Other plot: 
\code{\link{get_colour_palette}()},
\code{\link{get_xaxis_breaks}()}
}
\author{
Danielle Dempsey
}
\concept{plot}
