% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_deployment_log.R
\name{read_deployment_log}
\alias{read_deployment_log}
\title{Extract information from deployment log}
\usage{
read_deployment_log(path.log)
}
\arguments{
\item{path.log}{File path to the Log folder. T}
}
\value{
Returns a list with 5 elements. \code{deployment.dates} is a dataframe
 with two columns: \code{start.date} (the date of deployment) and
 \code{end.date} (date of retrieval). area.info is a dataframe with four
 columns: waterbody, latitude, longitude, and station. HOBO, aM, and vemco
 are each a dataframe with two columns: SENSOR (serial number) and the
 corresponding DEPTH (depth of deployment in m).
}
\description{
Extract information from deployment log
}
\details{
describe required columns here

 .xls or .xlsx files
}
\seealso{
Other compile: 
\code{\link{compile_HOBO_data}()},
\code{\link{compile_all_data}()},
\code{\link{compile_aquaMeasure_data}()},
\code{\link{compile_vemco_data}()}
}
\author{
Danielle Dempsey
}
\concept{compile}
