% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opendata_csv.R
\name{read_mooring_csv}
\alias{read_mooring_csv}
\title{Read mooring table}
\usage{
read_mooring_csv(path)
}
\arguments{
\item{path}{Path to mooring table (including file name and extension).}
}
\value{
Mooring table with standardized WATERBODY and STATION columns.
}
\description{
Read mooring table
}
