% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opendata_csv.R
\name{read_opendata_csv}
\alias{read_opendata_csv}
\title{Read in csv file of county data assembled for Open Data Portal}
\usage{
read_opendata_csv(path)
}
\arguments{
\item{path}{Path to the csv file, including file name and extension.}
}
\value{
Returns a dataframe of the assembled data.
}
\description{
Reads in the file using data.table::fread and assigns the
  appropriate class to each column.
}
