% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\docType{data}
\name{tidy_data}
\alias{tidy_data}
\title{Example data in tidy format}
\format{
This example data is a dataframe with 107898 rows and 6 variables:
  \describe{ \item{DEPLOYMENT_PERIOD}{Deployment and retrieval dates}
  \item{SENSOR}{Sensor type and serial number} \item{TIMESTAMP}{Timestamp of
  the observation} \item{VARIABLE}{VARIABLE MEASURED}
  \item{DEPTH}{Depth of the sensor} \item{VALUE}{Value of measurement} }
}
\source{
Data from Coastal Monitoring Program
}
\usage{
tidy_data
}
\description{
Example data in tidy format
}
\keyword{datasets}
