% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_data.R
\name{trim_data}
\alias{trim_data}
\title{Trim data to specified start and end datetimes}
\usage{
trim_data(
  dat.tidy,
  var.to.trim,
  start.datetime = min(dat.tidy$DATE),
  end.datetime = max(dat.tidy$DATE),
  sensors.to.trim = c("HOBO", "aquaMeasure", "VR2AR")
)
}
\arguments{
\item{dat.tidy}{Data in a tidy format, as exported by the
\code{convert_to_tidydata()} function.}

\item{var.to.trim}{Variable to trim, e.g. "Temperature", "Dissolved Oxygen",
or "Salinity".}

\item{start.datetime}{A character string indicating the date and time that the
sensor started reliably recording data, in one of the following orders: "ymd
IMS p", "Ymd IMS p", "Ymd HM", "Ymd HMS".}

\item{end.datetime}{A character string indicating the date and time the sensor
stopped reliably recording data, in one of the following orders: "ymd IMS
p", "Ymd IMS p", "Ymd HM", "Ymd HMS".}

\item{sensors.to.trim}{Sensor type(s) for which to trim the data. Default is
\code{sensors.to.trim = c("HOBO", "aquaMeasure", "VR2AR")}.}
}
\value{
Returns tidy.data trimmed to start.datetime to end.datetime
}
\description{
Trim data to specified start and end datetimes
}
\seealso{
Other format: 
\code{\link{convert_depth_to_ordered_factor}()},
\code{\link{convert_to_tidydata}()},
\code{\link{format_for_opendata}()}
}
\author{
Danielle Dempsey
}
\concept{format}
