% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_degree_days.R
\name{calculate_degree_days}
\alias{calculate_degree_days}
\title{Calculates temperature in degree-days}
\usage{
calculate_degree_days(
  dat.tidy,
  ...,
  start.date = min(dat.tidy$TIMESTAMP),
  end.date = max(dat.tidy$TIMESTAMP)
)
}
\arguments{
\item{dat.tidy}{Data in tidy format, as returned by the function
\code{convert_to_tidydata()}. Must include three columns: \code{TIMESTAMP}
(POSIXct), \code{VARIABLE} (character), and \code{VALUE} (numeric). May also
include columns with grouping variables passed to \code{...}. Other columns
will be ignored.}

\item{...}{Columns in \code{.dat.tidy} to use for grouping in
\code{dplyr::group_by()}, e.g., \code{YEAR, MONTH}. Degree-days are
calculated for each combination of groups.}

\item{start.date}{First day that should be included in the calculation. If
\code{start.date} does not include a time, then the start time is assumed
to be midnight. Accepted orders for \code{start.date} are: "ymd IMS p",
"Ymd IMS p", "Ymd HM", "Ymd HMS", "dmY HM", "dmY HMS", "Ymd", "ymd". Default
is the first day in \code{dat.tidy}.}

\item{end.date}{Last day that should be included in the calculation. If
\code{end.date} does not include a time, then the end time is assumed to be
"23:59:59". Accepted orders for \code{end.date} are: "ymd IMS p", "Ymd IMS
p", "Ymd HM", "Ymd HMS", "dmY HM", "dmY HMS", "Ymd", "ymd". Default is the
last day in \code{dat.tidy}.}
}
\value{
Returns a tibble with at least five columns: \code{PERIOD} (start and
 end date used to calculate mean temperature and number of days),
 \code{n_DAYS} (the number of days, \emph{n}, used in the calculation),
 \code{n_OBSERVATIONS} (the number of observations used to calculate the
 average temperature), \code{AVG_TEMPERATURE} (the average temperature in the
 time period), \code{DEGREE_DAYS} (degree-days, the product of
 \code{AVG_TEMPERATURE} and \code{n_DAYS}). Additional columns are returned
 for each grouping variable in \code{...}.
}
\description{
Calculates temperature in degree-days
}
\details{
Degree-days = average temperature over \emph{n} days * \emph{n}

 \emph{n} is calculated using the \code{difftime()} function:

 \code{n_DAYS = round(as.numeric(difftime(max(TIMESTAMP), min(TIMESTAMP),
 units = "days")), digits = 3)}
}
\examples{
data(tidydata)

# degree-days including all depths for whole time series
calculate_degree_days(tidydata)

# degree-days by DEPTH for whole time series
calculate_degree_days(tidydata, DEPTH)

# degree-days by DEPTH and month from July 1 to September 30
calculate_degree_days(tidydata, DEPTH, months(TIMESTAMP),
start.date = "2019-06-01", end.date = "2019-09-30")
}
\seealso{
Other calculate: 
\code{\link{calculate_DO_concentration}()}
}
\author{
Danielle Dempsey
}
\concept{calculate}
