% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DO_concentration.R
\name{calculate_DO_concentration}
\alias{calculate_DO_concentration}
\title{Calculates dissolved oxygen concentration (mg/L)}
\usage{
calculate_DO_concentration(
  dat.wide,
  Sal = NULL,
  P_atm = NULL,
  return.cp = FALSE
)
}
\arguments{
\item{dat.wide}{Dataframe with columns \code{TIMESTAMP}, \code{Temperature}
(degrees Celsius), and \code{DO_percent_sat} (\% saturation). Other columns
will be ignored.}

\item{Sal}{A single value of salinity (ppt). This value should be specified
if there is no \code{Salinity} column in \code{dat.wide}. Default is
\code{Sal = NULL}. Note: if \code{Sal} is specified when there is a
\code{Salinity} column in \code{dat.wide}, function will stop with an
error.}

\item{P_atm}{A single value of barometric pressure (atm). This value should
be specified if there is no \code{Pressure} column in \code{dat.wide}.
Default is \code{P_atm = NULL}. Note: if \code{P_atm} is specified when
there is a \code{Pressure} column in \code{dat.wide}, function will stop
with an error.}

\item{return.cp}{Logical parameter. If \code{TRUE}, the function returns a
  wide dataframe with columns: \code{TIMESTAMP}, \code{Temperature},
  \code{DO_concentraion} (in mg/L), the parameters used to calculate
  \code{C_p}, \code{C_p}, and \code{DO_percent_sat}.

  If \code{FALSE}, the function returns a wide dataframe with columns:
  \code{TIMESTAMP}, \code{Temperature}, \code{DO_concentration} (dissolved
  oxygen in units of mg/L).}
}
\description{
Converts dissolved oxygen from percent saturation to
 concentration (mg/L) based on temperature, salinity, and barometric pressure.
}
\details{
DO concentration is calculated as: Where C_p is the solubility of
 oxygen at the observed temperature, pressure, and salinity.

 C_p is calculated using equations 24 and 32, and Table 2 from Benson and
 Krause 1984.

 Benson, Bruce B., Krause, Daniel, (1984), The concentration and isotopic
 fractionation of oxygen dissolved in freshwater and seawater in equilibrium
 with the atmosphere, Limnology and Oceanography, 3, doi:
}
\seealso{
Other calculate: 
\code{\link{calculate_DO_percent_saturation}()}
}
\author{
Danielle Dempsey, Nicole Torrie
}
\concept{calculate}
