% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DO_concentration.R
\name{calculate_DO_concentration}
\alias{calculate_DO_concentration}
\title{Calculates dissolved oxygen concentration (mg/L)}
\usage{
calculate_DO_concentration(dat.tidy, pressure)
}
\arguments{
\item{dat.tidy}{Data in tidy format, as returned by the function
\code{convert_to_tidydata()}. Must include three columns: \code{SENSOR}
(character, must include "aquaMeasure-xxxxxx"), \code{VARIABLE} (character,
must include "Temperature" and "Dissolved Oxygen"), and \code{VALUE}
(numeric). Temperature values must be in degrees Celsius, and Dissolved
Oxygen values must be in \% saturation. Other columns will be ignored.}

\item{pressure}{Total pressure (water pressure + air pressure) in atm on the
DO sensor. See Details section for how to determine an appropriate value.}
}
\value{
Returns \code{dat.tidy} with additional observations in the
 \code{VARIABLE} column for "DO_concentration" and the corresponding value in
 the \code{VALUE} column.
}
\description{
Converts dissolved oxygen from percent saturation to
 concentration (mg/L) based on temperature and pressure.
}
\details{
DO concentration is calculated as:

 \deqn{DO_{mg/L} = C_p * DO_{\% Saturation} / 100}

 To see the calculation for \eqn{C_p}, go to
 \url{https://www.waterontheweb.org/under/waterquality/oxygen.html}.

 To determine what value to use for the \code{pressure} argument:

 \enumerate{

 \item Calculate water pressure: Navigate to:
 \url{https://bluerobotics.com/learn/pressure-depth-calculator/?waterType=fresh}.
  Choose "Freshwater" or "Saltwater" and use the calculator to determine the
 pressure of water above the DO sensor in atm.

 \item Calculate air pressure:

 If the body of water is at sea level, air pressure is 1 atm.

 If the body of water is not at sea level, first determine the altitude from:
 \url{https://www.mapcoordinates.net/en}. Next, determine the air pressure in
 atm from \url{https://www.mide.com/air-pressure-at-altitude-calculator}
 (using the "Calculate Air Pressure at Altitude" calculator).

 \item Total pressure = water pressure + air pressure. Use the total pressure
 value as the \code{pressure} argument. }
}
\seealso{
Other calculate: 
\code{\link{calculate_degree_days}()}
}
\author{
Danielle Dempsey, Nicole Torrie
}
\concept{calculate}
