% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DO_percent_saturation.R
\name{calculate_DO_percent_saturation}
\alias{calculate_DO_percent_saturation}
\title{Calculates dissolved oxygen percent saturation (\%)}
\usage{
calculate_DO_percent_saturation(
  dat.wide,
  air.pressure = 1,
  sensor.depth,
  salt.water = TRUE,
  return.cp = FALSE
)
}
\arguments{
\item{dat.wide}{Temperature values must be in degrees Celsius, and Dissolved
Oxygen values must be in mg/L. Other columns will be ignored and not
returned.}

\item{air.pressure}{Air pressure in atm on the DO sensor. See Details section
for how to determine an appropriate value.}

\item{sensor.depth}{Depth of the sensor from the surface in metres. Used to
calculate the water pressure on the sensor.}

\item{salt.water}{Logical parameter. If \code{TRUE} the density of salt water
is used to estimate water pressure. If \code{FALSE} the density of fresh
water is used to estimate water pressure. Default is \code{salt.water =
TRUE}.}

\item{return.cp}{Logical parameter. If \code{TRUE}, the function returns a
wide dataframe with columns \code{TIMESTAMP}, \code{Temperature},
\code{DO_percent_sat}, \code{DO_concentration}, and \code{cp}. If
\code{FALSE}, the function returns a wide dataframe with columns
\code{TIMESTAMP}, \code{Temperature}, \code{Dissolved Oxygen} (with
dissolved oxygen in units of percent saturation).}
}
\description{
Converts dissolved oxygen from concentration (mg/L) to percent
  saturation based on temperature and pressure.
}
\details{
DO percent saturation is calculated as:

  \deqn{DO_{\% Saturation} = 100 * DO_{mg/L} / C_p}


  To see the calculation for \eqn{C_p}, go to
  \url{https://www.waterontheweb.org/under/waterquality/oxygen.html}.

  To determine what value to use for the \code{pressure} argument:

  \enumerate{

  \item Calculate water pressure: Navigate to:
  \url{https://bluerobotics.com/learn/pressure-depth-calculator/?waterType=fresh}.
   Choose "Freshwater" or "Saltwater" and use the calculator to determine the
  pressure of water above the DO sensor in atm.

  \item Calculate air pressure:

  If the body of water is at sea level, air pressure is 1 atm.

  If the body of water is not at sea level, first determine the altitude
  from: \url{https://www.mapcoordinates.net/en}. Next, determine the air
  pressure in atm from
  \url{https://www.mide.com/air-pressure-at-altitude-calculator} (using the
  "Calculate Air Pressure at Altitude" calculator).

  \item Total pressure = water pressure + air pressure. }
}
\seealso{
Other calculate: 
\code{\link{calculate_DO_concentration}()}
}
\author{
Danielle Dempsey, Nicole Torrie
}
\concept{calculate}
