% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DO_percent_saturation.R
\name{calculate_DO_percent_saturation}
\alias{calculate_DO_percent_saturation}
\title{Calculates dissolved oxygen percent saturation (\%)}
\usage{
calculate_DO_percent_saturation(
  dat.wide,
  Sal = NULL,
  P_atm = NULL,
  return.cp = FALSE
)
}
\arguments{
\item{dat.wide}{Dataframe with columns \code{TIMESTAMP}, \code{Temperature}
(degrees Celsius), and \code{DO_concentration} (mg/L). Other columns will
be ignored.}

\item{Sal}{A single value of salinity (psu). This value must be specified if
there is no \code{Salinity} column in \code{dat.wide}. Default is \code{Sal
= NULL}. Note: if \code{Sal} is specified when there is a \code{Salinity}
column in \code{dat.wide}, the function will stop with an error.}

\item{P_atm}{A single value of barometric pressure (atm). This value should
be specified if there is no \code{Pressure} column in \code{dat.wide}.
Default is \code{P_atm = NULL}. Note: if \code{P_atm} is specified when
there is a \code{Pressure} column in \code{dat.wide}, function will stop
with an error.}

\item{return.cp}{Logical parameter. If \code{TRUE}, the function returns a
  wide dataframe with columns: \code{TIMESTAMP}, \code{Temperature},
  \code{DO_concentraion} (in mg/L), the parameters used to calculate
  \code{C_p}, \code{C_p}, and \code{DO_percent_sat}.

  If \code{FALSE}, the function returns a wide dataframe with columns:
  \code{TIMESTAMP}, \code{Temperature}, \code{DO_percent_sat} (dissolved
  oxygen in units of percent saturation).}
}
\description{
Converts dissolved oxygen from concentration (mg/L) to percent
  saturation based on temperature, salinity, and barometric pressure.
}
\details{
DO percent saturation is calculated as:

  \deqn{DO_{\% Saturation} = 100 * DO_{mg/L} / C_{p}}

  DO_{mg/L} is the concentration of dissolved oxygen in mg/L (make sure the
  salinity and/or pressure-corrected concentration is used if required).

  C_p is the solubility of oxygen at the observed temperature, pressure, and
  salinity.

  C_p is calculated using equations 24 and 32, and Table 2 from Benson and
  Krause 1984.

  Benson, Bruce B., Krause, Daniel, (1984), The concentration and isotopic
  fractionation of oxygen dissolved in freshwater and seawater in equilibrium
  with the atmosphere, Limnology and Oceanography, 3, doi:
  10.4319/lo.1984.29.3.0620.
}
\seealso{
Other calculate: 
\code{\link{calculate_DO_concentration}()}
}
\author{
Danielle Dempsey, Nicole Torrie
}
\concept{calculate}
