% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_deployment_logs.R
\name{combine_deployment_logs}
\alias{combine_deployment_logs}
\title{Combines logs from a station to make deployment history file}
\usage{
combine_deployment_logs(path)
}
\arguments{
\item{path}{Path to station folder.}
}
\value{
Exports the deployment history file to the station folder, in .csv
 format.
}
\description{
Combines logs from a station to make deployment history file
}
\details{
Imports the log from each deployment folder, binds them together, and
 re-formats into the deployment history file. Logs can be in .xls, .xlsx, or
 .csv format.
}
\author{
Danielle Dempsey
}
