% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_all_data.R
\name{compile_all_data}
\alias{compile_all_data}
\title{Compiles HOBO, aquaMeasure, and Vemco data from a single deployment}
\usage{
compile_all_data(
  path,
  area.name,
  deployment.range,
  trim,
  serial.table.HOBO = NULL,
  file.type.HOBO = NULL,
  serial.table.aM = NULL,
  depth.vemco = NULL
)
}
\arguments{
\item{path}{File path to the Hobo, aquaMeasure, and/or Vemco folders.}

\item{area.name}{Area where the sensor string was deployed.}

\item{deployment.range}{A dataframe with two columns. The first column holds
the deployment date (a Date object in the order year, month, day),  and the
second column holds the retrieval date (a Date object in the order year,
month, day).}

\item{trim}{Logical value indicating whether to trim the data to the dates
specified in \code{deployment.range}. (Note: four hours are added to the
retrieval date to account for AST, e.g., in case the sensor was retrieved
after 20:00 AST, which is 00:00 UTC the next day.) Default is \code{trim =
TRUE}.}

\item{serial.table.HOBO}{A table with the serial number of each HOBO and
TidBiT sensor on the string, in the form "HOBO-xxxxxxxx" or
"TidbiT-xxxxxxxx" (first column) and corresponding depth at which it was
deployed in the form "2m" (second column).}

\item{file.type.HOBO}{Character string indicating whether the HOBO data is in
 .csv or .xlsx format. All HOBO data files that are being compiled must have
 the same file extension.  Default is \code{file.type.HOBO = NULL}. Alternatives
 are \code{file.type = "csv"} and \code{file.type = "xlsx"}.

 Note that the HOBO software exports csv files in true UTC, but xlsx files
 account for daylight savings time. This function will automatically convert
 the datetimes in xlsx files to true UTC. Future versions of the package will
 provide an option to convert from ADT to AST.}

\item{serial.table.aM}{A table with the serial number of each aquaMeasure on
the string, in the form "aquaMeasure-xxxxxx" (first column; note the capital
"M") and its corresponding depth in the form "2m" (second column).}

\item{depth.vemco}{Character string indicating the depth at which the Vemco
was deployed, in the format "10m".}
}
\value{
Returns a dataframe with the compiled and formatted data. Columns
 alternate between the timestamp (in the format "Y-m-d H:M:S") and the
 variable value (rounded to three decimal places). Metadata at the top of
 each column indicates the deployment dates, the sensor serial number, and
 the variable and depth of the sensor. Each datetime column shows the
 timezone as extracted from the sensor.

 To include the metadata, all values were converted to class
 \code{character}. To manipulate the data, the values must be converted to
 the appropriate class (e.g., \code{POSIXct} for the datetimes and
 \code{numeric} for variable values). This can be done using the function
 \code{convert_to_tidydata()}.
}
\description{
Compiles HOBO, aquaMeasure, and Vemco data from a single deployment
}
\details{
Calls \code{compile_HOBO_data()}, \code{compile_aquaMeasure_data()}
 and \code{compile_vemco_data()} and joins the results into a single
 dataframe.

 HOBO data must be in a folder named Hobo, aquaMeasure data must be in a
 folder named aquaMeasure, and Vemco data must be in a folder name Vemco. The
 Hobo, aquaMeasure, and Vemco folders must be in the same folder.

 If one type of sensor was not included in the deployment, set the
 argument(s) related to that sensor to \code{NULL} (the default). For
 example, if there is no aquaMeaure data for the deployment, set
 \code{serial.table.aM = NULL}
}
\seealso{
Other compile: 
\code{\link{compile_HOBO_data}()},
\code{\link{compile_aquaMeasure_data}()},
\code{\link{compile_vemco_data}()},
\code{\link{read_deployment_log}()}
}
\author{
Danielle Dempsey
}
\concept{compile}
