% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_aquaMeasure_data.R
\name{compile_aquaMeasure_data}
\alias{compile_aquaMeasure_data}
\title{Compiles temperature, dissolved oxygen, and salinity data from
 aquaMeasure deployment}
\usage{
compile_aquaMeasure_data(
  path.aM,
  area.name,
  vars.aM = c("Temperature", "Dissolved Oxygen", "Salinity"),
  depth.aM,
  deployment.range
)
}
\arguments{
\item{path.aM}{File path to the aquaMeasure data. Should end with
\code{/aquaMeasure}.}

\item{area.name}{Area where aquaMeasure was deployed.}

\item{vars.aM}{The variables to extract. Could possibly replace with
unique(Record Type)}

\item{depth.aM}{The depth at which the sensor was deployed.}

\item{deployment.range}{The start and end dates of deployment from the
deployment log. Must be in format "2018-Nov-15 to 2020-Jan-24".}
}
\value{
Returns a dataframe and exports a spreadsheet with the aquaMeasure
 data, including the appropriate metadata. Note that to include the metadata,
 all values were converted to class \code{character}. To manipulate the data,
 the values must be converted to the appropriate class (e.g., \code{POSIXct}
 for \code{DATE}, \code{numeric} for temperature values). This can be done
 using the function \code{convert_to_tidydata()}.
}
\description{
This functions re-formats the data from an aquaMeasure deployment
 so it can be combined with the HOBO temperature data.
}
\details{
Replaces negative DO values with \code{NA}.
I haven't done anything with the DATES. Still need to un-account for
 daylight savings. To be discussed.
}
\seealso{
Other compile: 
\code{\link{compile_HOBO_data}()},
\code{\link{convert_to_tidydata}()}
}
\author{
Danielle Dempsey
}
\concept{compile}
