% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_HOBO_data.R
\name{compile_HOBO_data}
\alias{compile_HOBO_data}
\title{Compiles temperature data from HOBO and TidbiT sensors}
\usage{
compile_HOBO_data(
  path.HOBO,
  area.name = "",
  serial.table.HOBO,
  deployment.range,
  trim = TRUE,
  rm.DST = FALSE,
  export.csv = FALSE
)
}
\arguments{
\item{path.HOBO}{File path to the Hobo folder. All of the excel files in the
Hobo folder will be compiled. The name of each file must be the serial
number of the sensor, and the excel files must be in either .csv or .xlsx
format. The timestamp columns must be in the order "ymd IMS p", "Ymd IMS p",
"Ymd HM", "Ymd HMS", "dmY HM", or "dmY HMS".}

\item{area.name}{Area where the sensor(s) was deployed.}

\item{serial.table.HOBO}{A table with the serial number of each HOBO and
TidBiT sensor on the string, in the form "HOBO-xxxxxxxx" or
"TidbiT-xxxxxxxx" (first column) and corresponding depth at which it was
deployed in the form "2m" (second column).}

\item{deployment.range}{A dataframe with two columns. The first column holds
the deployment date (a Date object in the order year, month, day),  and the
second column holds the retrieval date (a Date object in the order year,
month, day).}

\item{trim}{Logical value indicating whether to trim the data to the dates
specified in \code{deployment.range}. (Note: four hours are added to the
retrieval date to account for AST, e.g., in case the sensor was retrieved
after 20:00 AST, which is 00:00 UTC the next day.) Default is \code{trim =
TRUE}.}

\item{rm.DST}{Option to remove daylight savings time. Here, daylight savings
is as defined for Nova Scotia, Canada (begins in March and end in November).
Only works for 2015 - 2021. It is strongly recommended that the data is in
the correct timezone before using this function, and that the default
\code{rm.DST = FALSE} is used.}

\item{export.csv}{Logical value indicating whether to export the compiled data
as a .csv file. If \code{export.csv = TRUE}, the compiled data will not be
returned to the global environment. Default is \code{export.csv = FALSE}.}
}
\value{
Returns a dataframe or exports a spreadsheet with the data compiled
 from each of the HOBO and TidbiT sensors. Columns alternate between
 timestamp (in the format "Y-m-d H:M:S") and temperature value (rounded to
 three decimal places). Metadata at the top of each column indicates the
 deployment and retrieval dates, the sensor serial number, the depth of the
 sensor, the temperature units, and the timezone of the timestamp.

 To include the metadata, all values were converted to class
 \code{character}. To manipulate the data, the values must be converted to
 the appropriate class (e.g., \code{POSIXct} for the timestamps and
 \code{numeric} for temperature values). This can be done using the function
 \code{convert_to_tidydata()}.
}
\description{
Compiles and formats data from HOBO and TidbiT sensors.
}
\details{
The raw HOBO and TidBiT data must be saved in a folder named Hobo in
 .csv or .xlsx format.

 All columns are read in as characters to ensure the timestamp is parsed
 correctly. Timestamp must be saved in Excel as a number or a character in
 the order ""ymd IMS p", "Ymd IMS p", "Ymd HM", "Ymd HMS", "dmY HM", or "dmY
 HMS".

 CMAR NOTES: Data should be exported from the Hobo software in GMT+00 as a
 csv file so that the timestamp is in UTC.

 If exported as a .xlsx file, the timestamp accounts for daylight savings
 time (this seems to be a bug in the HOBO software).
 \code{compile_HOBO_data()} can convert .xlsx files to true UTC by setting
 \code{rm.DST = TRUE}. This will subtract 1 hour from each datetime that
 occurs during daylight savings.

 The functions used to convert to true UTC are
 \code{convert_HOBO_datetime_to_true_UTC()} and \code{dates_to_fix()}, which
 are NOT exported to the \code{strings} package (i.e., they are only used
 internally).

 \code{convert_HOBO_datetime_to_true_UTC()} identifies which datetimes are
 during daylight savings by creating two additional columns: \code{ADT_force
 = force_tz(TIMESTAMP, tzone = "America/Halifax")}, and
 \code{DAYLIGHT_SAVINGS = dst(ADT_force)}. Where \code{DAYLIGHT_SAVINGS ==
 TRUE}, the \code{TIMESTAMP} is shifted back by 1 hour.

 This leaves apparent duplicates for the hour of 1 am on the day that
 daylight savings ends. \code{dates_to_fix()} identifies these \code{n}
 observations (e.g. 1:00, 1:15, 1:30, 1:45, 1:00, 1:15, 1:30, 1:45), and
 shifts the first \code{n/2} back by one hour (e.g. 00:00, 00:15, 00:30,
 00:45, 1:00, 1:15, 1:30, 1:45).The function is hard-coded 2015 - 2021.

 If for some reason, there is an ODD number of duplicates, the function might
 break.
}
\seealso{
Other compile: 
\code{\link{compile_all_data}()},
\code{\link{compile_aquaMeasure_data}()},
\code{\link{compile_vemco_data}()},
\code{\link{read_deployment_log}()}
}
\author{
Danielle Dempsey
}
\concept{compile}
