% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_HOBO_data.R
\name{compile_HOBO_data}
\alias{compile_HOBO_data}
\title{Compiles temperature data from HOBO deployment}
\usage{
compile_HOBO_data(
  path.HOBO,
  area.name,
  serial.table,
  deployment.range,
  trim = TRUE,
  file.type = "csv",
  export.csv = FALSE
)
}
\arguments{
\item{path.HOBO}{File path to the Hobo folder. All of the excel files in the
Hobo folder should be data extracted from the HOBO software. The name of
each excel file must be the serial number of the sensor, and the excel files
must all have the same extension (either .csv or .xlsx). The datetime
columns must be in the order "ymd IMS p", "Ymd HM", or "Ymd HMS".}

\item{area.name}{Area where the HOBO was deployed.}

\item{serial.table}{A table with the serial number of each HOBO in the form
"HOBO-xxxxxxxx" (first column) and corresponding variable it measured at the
depth it was deployed in the form "Temperature-2m" (second column).}

\item{deployment.range}{The start and end dates of deployment from the
deployment log. Must be in format "2018-Nov-15 to 2020-Jan-24".}

\item{trim}{Logical value indicating whether to trim the data to the dates
specified in \code{deployment.range}. (Note: four hours are added to the
retrieval date to account for AST, e.g., in case the sensor was retrieved
after 20:00 AST, which is 00:00 UTC the next day.) Default is \code{trim =
TRUE}.}

\item{file.type}{Character string indicating whether the HOBO data is in .csv
or .xlsx format. All data files that are being compiled must have the same
file extension. Note that the HOBO software exports csv files in true UTC,
but xlsx files account for daylight savings time. This function will
automatically convert the datetimes in xlsx files to true UTC. Default is
\code{file.type = "csv"}. Alternative is \code{file.type = "xlsx"}.}

\item{export.csv}{Logical value indicating whether to export the compiled data
as a .csv file. If \code{export.csv = TRUE}, the compiled data will not be
returned to the global environment. Default is \code{export.csv = FALSE}.}
}
\value{
Returns a dataframe or exports a spreadsheet with the data compiled
 from each of the HOBO sensors. Columns alternate between datetime (UTC, in
 the format "Y-m-d H:M:S") and temperature value (degree celsius, rounded to
 three decimal places). Metadata at the top of each column indicates the
 deployment range, the sensor serial number, and the depth of the sensor.
 Each datetime column shows the timezone as extracted from the HOBOware, and
 each temperature column shows the units extracted from HOBO.

 To include the metadata, all values were converted to class
 \code{character}. To manipulate the data, the values must be converted to
 the appropriate class (e.g., \code{POSIXct} for the datetimes and
 \code{numeric} for temperature values). This can be done using the function
 \code{convert_to_tidydata()}.
}
\description{
This function compiles the data from a HOBO deployment at
 different depths into a single dataframe or spreadsheet.
}
\details{
All columns are read in as characters to ensure the timestamp is
 parsed correctly.

 Can handle .csv or .xlsx files, but .csv files are preferred (see below).

 HOBO data should be exported in GMT+00 as a csv file so that the timestamp
 is in UTC.

 If exported as an xlsx, the timestamp accounts for daylight savings time.
 \code{compile_HOBO_data()} will automatically convert xlsx files to true UTC
 by subtracting 1 hour from each datetime that occurs during daylight
 savings. This feature will be removed as we solidify our workflow.

 The functions used to convert to true UTC are
 \code{convert_HOBO_datetime_to_true_UTC()} and \code{dates_to_fix()}, which
 are NOT exported to the \code{strings} package (i.e., they are only used
 internally).

 \code{convert_HOBO_datetime_to_true_UTC()} identifies which times are during
 daylight savings by creating two additional columns: \code{ADT_force =
 force_tz(DATE, tzone = "America/Halifax")}, and \code{DAYLIGHT_SAVINGS =
 dst(ADT_force)}. Where \code{DAYLIGHT_SAVINGS == TRUE}, the \code{DATE} is
 shifted back by 1 hour.

 This leaves apparent duplicates for the hour of 1 am on the day that
 daylight savings ends. \code{dates_to_fix()} identifies these \code{n}
 observations (e.g. 1:00, 1:15, 1:30, 1:45, 1:00, 1:15, 1:30, 1:45), and
 shifts the first \code{n/2} back by one hour (e.g. 00:00, 00:15, 00:30,
 00:45, 1:00, 1:15, 1:30, 1:45). Function can handing this from 2017 - 2021;
 additional lines required for other years.

 If for some reason, there is an ODD number of duplicates, the function might
 break.
}
\seealso{
Other compile: 
\code{\link{compile_all_data}()},
\code{\link{compile_aquaMeasure_data}()},
\code{\link{compile_vemco_data}()},
\code{\link{convert_to_tidydata}()}
}
\author{
Danielle Dempsey
}
\concept{compile}
