% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_DO_percent_to_concentration.R
\name{convert_DO_percent_to_concentration}
\alias{convert_DO_percent_to_concentration}
\title{Converts dissolved oxygen from percent to concentration (mg/L)}
\usage{
convert_DO_percent_to_concentration(dat.tidy, pressure)
}
\arguments{
\item{dat.tidy}{Tidy data **fill in here}

\item{pressure}{Nicole to send blurb}
}
\value{
Fill this in. Returns a tidy dataframe with six columns: \code{DATE} (POSIXct),
 \code{VALUE}, (numeric), \code{DATE_RANGE} (character), \code{SENSOR},
 \code{VARIABLE} (character), \code{DEPTH} (ordered factor).
}
\description{
Converts dissolved oxygen from percent to concentration (mg/L)
}
\seealso{
Other format: 
\code{\link{convert_depth_to_ordered_factor}()},
\code{\link{convert_to_tidydata}()},
\code{\link{format_for_opendata}()},
\code{\link{trim_data}()}
}
\author{
Danielle Dempsey
}
\concept{format}
