% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_tidydata.R
\name{convert_to_tidydata}
\alias{convert_to_tidydata}
\title{Converts compiled string data to a tidy format}
\usage{
convert_to_tidydata(dat.wide, remove.NA = TRUE, show.NA.message = FALSE)
}
\arguments{
\item{dat.wide}{Data in the format exported by the \code{compile_xxx_data}
functions.}

\item{remove.NA}{Logical value. If \code{remove.NA = TRUE}, observations where
\code{DATE} or \code{VALUE} = \code{NA} are removed. Default is
\code{remove.NA = TRUE}.}

\item{show.NA.message}{Logical value. If \code{show.NA.message = TRUE}, the
number of NA values removed is printed to the console. Default is
\code{show.NA.message = FALSE}.}
}
\value{
Returns a tidy dataframe with six columns: \code{DATE} (POSIXct),
 \code{VALUE}, (numeric), \code{DATE_RANGE} (character), \code{SENSOR},
 \code{VARIABLE} (character), \code{DEPTH} (ordered factor).
}
\description{
Converts compiled string data to a tidy format
}
\seealso{
Other format: 
\code{\link{format_for_opendata}()}
}
\author{
Danielle Dempsey
}
\concept{format}
