% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_for_opendata.R
\name{format_for_opendata}
\alias{format_for_opendata}
\title{Format sensor string data for the OpenData portal}
\usage{
format_for_opendata(dat.tidy, location.info)
}
\arguments{
\item{dat.tidy}{Data from a single deployment in tidy format, as returned by
the function \code{convert_to_tidydata()}. Must include four columns:
\code{DATE} (POSIXct), \code{VARIABLE} (character), \code{DEPTH} (ordered
factor), and \code{VALUE} (numeric).}

\item{location.info}{Dataframe of information about the deployment location. One
observation of 5 columns: \code{county}, \code{waterbody}, \code{station},
\code{latitude}, and \code{longitude}.}
}
\value{
Returns a dataframe of tidy data in the format for the OpenData portal for a single deployment,
}
\description{
Format sensor string data for the OpenData portal
}
\seealso{
Other format: 
\code{\link{convert_DO_percent_to_concentration}()},
\code{\link{convert_depth_to_ordered_factor}()},
\code{\link{convert_to_tidydata}()},
\code{\link{trim_data}()}
}
\author{
Danielle Dempsey
}
\concept{format}
