% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot.R
\name{get_xaxis_breaks}
\alias{get_xaxis_breaks}
\title{Returns nice major and minor breaks and label format based on timespan
 of the data}
\usage{
get_xaxis_breaks(timespan)
}
\arguments{
\item{timespan}{Number of days of data to be plotted.}
}
\value{
Returns a dataframe with 1 observation of 3 variables
 \code{date.breaks.major}, \code{date.breaks.minor},
 \code{date.labels.format}.
}
\description{
Returns nice major and minor breaks and label format based on timespan
 of the data
}
\details{
If \code{timespan} is less than or equal to 60 days, the major and
 minor breaks will be every 2 weeks; if time span is greater than 60 days but
 less than or equal to 120 days, the major and minor breaks will be every
 month; if time span is greater than 120 days, the major breaks will be 2
 months and the minor breaks will be 1 month.
}
\seealso{
Other OpenData: 
\code{\link{write_report_table}()}
}
\author{
Danielle Dempsey
}
\concept{OpenData}
