% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot.R
\name{get_xaxis_breaks}
\alias{get_xaxis_breaks}
\title{Returns nice major and minor breaks and label format based on timespan
 of the data}
\usage{
get_xaxis_breaks(dat.tidy)
}
\arguments{
\item{dat.tidy}{Data to be plotted, in tidy format, as returned by the
function \code{convert_to_tidydata()}. Must include the column
\code{TIMESTAMP} (POSIXct).}
}
\value{
Returns a dataframe with 1 observation of 3 variables
 \code{date.breaks.major}, \code{date.breaks.minor},
 \code{date.labels.format}.
}
\description{
Returns nice major and minor breaks and label format based on timespan
 of the data
}
\details{
If the timespan in the \code{TIMESTAMP} column of \code{dat.tidy} is
 less than or equal to 60 days, the major and minor breaks will be every 2
 weeks; if time span is greater than 60 days and less than or equal to 120
 days, the major and minor breaks will be every month; if time span is
 greater than 120 days and less than or equal to 240 days, the major breaks
 will be 2 months and the minor breaks will be 1 month; if the time span is
 greater than 240 days, the major breaks will be 4 months and the minor
 breaks will be 1 month.
}
\seealso{
Other plot: 
\code{\link{plot_variables_at_depth}()},
\code{\link{set_colour_palette}()}
}
\author{
Danielle Dempsey
}
\concept{plot}
