% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_variables_at_depth.R
\name{ggplot_variables_at_depth}
\alias{ggplot_variables_at_depth}
\title{Plot variables at depth with faceted ggplot}
\usage{
ggplot_variables_at_depth(
  dat.tidy,
  superchill = FALSE,
  heatstress = FALSE,
  trending_up
)
}
\arguments{
\item{dat.tidy}{Tidy data}

\item{superchill}{Logical}

\item{heatstress}{Logical}

\item{trending_up}{Logical}
}
\value{
ggplot object
}
\description{
Plot variables at depth with faceted ggplot
}
