% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_compiled_data.R
\name{name_compiled_data}
\alias{name_compiled_data}
\title{Generates name for output file of compiled string data}
\usage{
name_compiled_data(area.name, deployment.start, vars)
}
\arguments{
\item{area.name}{Area where the sensor(s) was deployed.}

\item{deployment.start}{Start date of deployment}

\item{vars}{The variables included in the data file, e.g. \code{vars =
c("Temperature", "Dissolved Oxygen")}.}
}
\description{
Generates a name for output file in the format "area
 name_deployment date_variables", e.g. "Shad Bay_2018-11-15_TEMP_DO".
}
