% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variables_at_depth.R
\name{plot_variables_at_depth}
\alias{plot_variables_at_depth}
\title{Exports ggplot2 objects for temperauture, dissolved oxygen, and
 salinity on separate panes}
\usage{
plot_variables_at_depth(
  dat.tidy,
  plot.title = "",
  vars.to.plot = c("Temperature", "Dissolved Oxygen", "Salinity"),
  ylab.units = c(expression(paste("(", degree, "C)")), "(\%)", "(ppt)"),
  color.palette = rev(viridis(6, option = "D")),
  date.breaks.major = "2 month",
  date.breaks.minor = "1 month",
  stacked = TRUE
)
}
\arguments{
\item{dat.tidy}{Data in tidy format, as returned by the function
\code{convert_to_tidydata()}. Must include four columns: \code{DATE}
(POSIXct), \code{VALUE}, (numeric), \code{VARIABLE} (character),
\code{DEPTH} (ordered factor).}

\item{plot.title}{Title for plot. Default "" (will this work for loop?)}

\item{vars.to.plot}{A character vector of variables for which to return plots.
Options are any combination of "Temperature", "Dissolved Oxygen", and
"Salinity". Default is \code{vars.to.plot = c("Temperature", "Dissolved
Oxygen", "Salinity")}.}

\item{ylab.units}{Character vector of units. Must be in correct order!}

\item{color.palette}{Color palette of hex colors onto which \code{DEPTH} will
be mapped. Default is \code{color.palette = rev(viridis(6, option = "D"))}.}

\item{date.breaks.major}{Intervals for major breaks. Default is
\code{date.breaks.major = "2 month"}.}

\item{date.breaks.minor}{Intervals for minor breaks. Default is
\code{date.breaks.minor = "1 month"}.}

\item{stacked}{If true;  returns 1 figure; if FALSE, returns each separately}
}
\value{
Returns a list of ggplot2 plots of Temperature, Dissolved Oxygen,
 and/or Salinity.
}
\description{
Exports ggplot2 objects for temperauture, dissolved oxygen, and
 salinity on separate panes
}
\details{
can only do 4 vars
}
\author{
Danielle Dempsey
}
\concept{plot}
