% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variables_at_depth.R
\name{plot_variables_at_depth}
\alias{plot_variables_at_depth}
\title{Exports ggplot2 object(s) of variables at depth over time}
\usage{
plot_variables_at_depth(
  dat.tidy,
  plot.title = "",
  vars.to.plot = c("Temperature", "Dissolved Oxygen", "Salinity"),
  ylab.units = c("deg C)", "(\%)", "(ppt)"),
  color.palette = rev(viridis(6, option = "D")),
  date.breaks.major = "2 month",
  date.breaks.minor = "1 month",
  alpha.points = 1,
  stacked = TRUE
)
}
\arguments{
\item{dat.tidy}{Data in tidy format, as returned by the function
\code{convert_to_tidydata()}. Must include four columns: \code{DATE}
(POSIXct), \code{VARIABLE} (character), \code{DEPTH} (ordered factor), and
\code{VALUE} (numeric).}

\item{plot.title}{Title for plot. Default is blank: \code{plot.title = ""}.}

\item{vars.to.plot}{A character vector of variables to plot. Each string must
match an entry in the \code{VARIABLE} column of \code{dat.tidy}. Default is
\code{vars.to.plot = c("Temperature", "Dissolved Oxygen", "Salinity")}.}

\item{ylab.units}{Character vector of units as they will appear in the y-axis
title. Must be in the same order as the associated variable in
\code{vars.to.plot}. Note: the units for \code{vars.to.plot = "Temperature"}
are hard-coded to ensure the degree symbol renders.}

\item{color.palette}{Color palette of hex colors onto which \code{DEPTH} will
be mapped. Default is \code{color.palette = rev(viridis(6, option = "D"))}.}

\item{date.breaks.major}{Intervals for major breaks. Default is
\code{date.breaks.major = "2 month"}.}

\item{date.breaks.minor}{Intervals for minor breaks. Default is
\code{date.breaks.minor = "1 month"}.}

\item{alpha.points}{Value indicating the transparency of the points. 0 is most
transparent; 1 is opaque.}

\item{stacked}{Logical value indicating what figures to return.  If
\code{stacked = TRUE}, a single figure with the plots for each variable in
\code{vars.to.plot} stacked in a column is returned. Plots will be stacked
in the order the variables appear in \code{vars.to.plot}. \code{plot.title}
will be placed on the top plot. If \code{stacked = FALSE}, a list of the
individual plot for each variable will be returned. Default is \code{stacked
= TRUE}.}
}
\value{
Returns ggplot2 object(s). If \code{stacked = TRUE}, a single figure
 with the plots for each variable in \code{vars.to.plot} stacked in a column
 is returned. Plots will be stacked in the order the variables appear in
 \code{vars.to.plot}. \code{plot.title} will be placed on the top plot. If
 \code{stacked = FALSE}, a list of the individual plot for each variable will
 be returned.
}
\description{
Exports ggplot2 object(s) of variables at depth over time
}
\author{
Danielle Dempsey
}
\concept{plot}
