% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_deployment_log.R
\name{read_deployment_log}
\alias{read_deployment_log}
\title{Extract information from deployment log}
\usage{
read_deployment_log(path.log)
}
\arguments{
\item{path.log}{File path to the Log folder.}
}
\value{
Returns a list with 5 elements. \code{deployment.dates} is a dataframe
 with two columns: \code{start.date} (the date of deployment) and
 \code{end.date} (date of retrieval). area.info is a dataframe with five
 columns: waterbody, latitude, longitude, station, and lease. HOBO, aM, and
 vemco are each a dataframe with two columns: SENSOR (serial number) and the
 corresponding DEPTH (depth of deployment in m).
}
\description{
Extract information from deployment log
}
\details{
The log must be saved in .csv, .xlsx or .xls format, and must include
 the following columns:

 \code{Deployment_Waterbody} (waterbody where string was deployed),
 \code{Location_Description} (the station name), \code{Lease#} If located on
 an aquaculture site, the lease number (NA otherwise),
 \code{Deployment} (the deployment date, in the order "Ymd"),
 \code{Retrieval} (the retrieval date, in the order "Ymd"),
 \code{Logger_Latitude} (the latitude at which the string was deployed),
 \code{Logger_Longitude} (the longitude at which the string was deployed),
 \code{Logger_Model} (the type of sensor; see below for options),
 \code{Serial#} (the sensor serial number), and \code{Sensor_Depth} (depth at
 which the sensor was deployed). All other columns will be ignored.

 Entries in the \code{Logger_Model} column can be "HOBO Pro V2", "TidbiT
 MX2303"  "aquaMeasure DOT", "aquaMeasure SAL", "aquaMeasure SST", or
 "VR2AR". (Some mis-spellings are accepted: ("HOBO pro V2", "HOBO_Pro_V2",
 "aquameasure DOT", "aquameasure SAL", "aquameasure SST")

 A Warning message is printed to the console when the function does not
 recognize a sensor in the log.

 A message is printed to the console when Hobo, aquaMeasure, or Vemco sensors
 are not found in the log.

 A message will be printed to the console if there is more than one unique
 entry in \code{Deployment_Waterbody}, \code{Location_Description},
 \code{Deployment}, \code{Retrieval}, \code{Logger_Latitude}, or
 \code{Logger_Longitude}.
}
\seealso{
Other compile: 
\code{\link{compile_HOBO_data}()},
\code{\link{compile_all_data}()},
\code{\link{compile_aquaMeasure_data}()},
\code{\link{compile_vemco_data}()}
}
\author{
Danielle Dempsey
}
\concept{compile}
